-blogs page
- functionality on blogs page broke,
was considering removing it regardless.
never kept up with it, and it was a bit personal and vent-ey.
files still in the repo for archival sake, but the blogs.html page is non-functional and not linked anymore.
direct links to blog posts should still function, though, why would anyone need them.